# Change eth-hash backend to pysha3