### What was wrong?
PyCryptodome is horrifically slow and eth-hash was using it as a backend.
@jmyles laptop was getting too hot for his lap.


### How was it fixed?
Changed the eth-hash backend to use pysha3/hashlib.
@jmyles laptop cooled down.


#### Cute Animal Picture

![this cute ass penguin photo](https://wallpapercave.com/wp/u2FZQby.jpg)