@pipermerriam 2018-06-18
Also, since this is a performance item, would you mind posting the before/after benchmarking results from the benchmarking test suite.  Runnable as `python scripts/benchmark/run.py`