@tuxxy 2018-06-18
Sorry, I'm traveling. I'll have it fixed here in sec.

On Mon, Jun 18, 2018, 16:03 Jason Carver <notifications@github.com> wrote:

> *@carver* commented on this pull request.
> ------------------------------
>
> In setup.py
> <https://github.com/ethereum/py-evm/pull/921#discussion_r196251590>:
>
> > @@ -26,7 +27,8 @@
>      'evm-extra': [
>          "coincurve>=7.0.0,<8.0.0",
>          "plyvel==1.0.4",
> -        "eth-hash[pycryptodome]",
> +        "eth-hash[pysha3];implementation_name=='cpython'",
> +        "eth-hash[pycryptodome];implementation_nam=='pypy'",
>
> s/nam/name/
>
> —
> You are receiving this because you authored the thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/ethereum/py-evm/pull/921#pullrequestreview-129769293>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/ADbn7rajqn7lxxz417-pqssEdEIP5hAxks5t-DHWgaJpZM4UqlJc>
> .
>