@carver 2018-06-18
FWIW, you could theoretically override this locally without a code change to `trinity`: by installing `pysha3` and setting the environment variable to select the `pysha3` backend for eth-hash: http://eth-hash.readthedocs.io/en/stable/quickstart.html#specify-backend-by-environment-variable

```sh
$ ETH_HASH_BACKEND="pysha3" trinity ...
```