@tuxxy 2018-06-19
Note: tests were increasing with speed after each run. Posted is the worst pycryptodome run and the best pysha3 run. I couldn't really see a huge improvement, but it seems that pycryptodome was running slower on average. It was much easier for pysha3 to run faster than pycryptodome, it seems. But, the results aren't very conclusive. I'm not certain if this is due to how py-evm does benchmarking.

Comparing run times of pycryptodome vs pysha3 libs independent of py-evm shows very conclusive results that pysha3 is exceptionally faster than pycryptodome.

Benchmark before change:
```

______                 _                          _     
| ___ \               | |                        | |    
| |_/ / ___ _ __   ___| |__  _ __ ___   __ _ _ __| | __ 
| ___ \/ _ \ '_ \ / __| '_ \| '_ ` _ \ / _` | '__| |/ / 
| |_/ /  __/ | | | (__| | | | | | | | | (_| | |  |   <  
\____/ \___|_| |_|\___|_| |_|_| |_| |_|\__,_|_|  |_|\_\

Py-EVM/0.2.0a26/linux/cpython3.6.5

Starting benchmark: Empty block mining

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     7.744      |       0        |     0.000      |      500       |       64.563       |       0        |     0.000      |
|     homestead     |     7.907      |       0        |     0.000      |      500       |       63.238       |       0        |     0.000      |
| tangerine-whistle |     7.963      |       0        |     0.000      |      500       |       62.788       |       0        |     0.000      |
|  spurious-dragon  |     8.279      |       0        |     0.000      |      500       |       60.393       |       0        |     0.000      |
|     byzantium     |     8.795      |       0        |     0.000      |      500       |       56.847       |       0        |     0.000      |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     40.689     |       0        |       -        |      2500      |         -          |       0        |       -        |
|        Avg        |     8.138      |       0        |     0.000      |      500       |       61.442       |       0        |     0.000      |
================================================================================================================================================

Starting benchmark: Empty block import

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     8.798      |       0        |     0.000      |      500       |       56.832       |       0        |     0.000      |
|     homestead     |     8.767      |       0        |     0.000      |      500       |       57.033       |       0        |     0.000      |
| tangerine-whistle |     8.825      |       0        |     0.000      |      500       |       56.658       |       0        |     0.000      |
|  spurious-dragon  |     8.891      |       0        |     0.000      |      500       |       56.236       |       0        |     0.000      |
|     byzantium     |     9.193      |       0        |     0.000      |      500       |       54.389       |       0        |     0.000      |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     44.474     |       0        |       -        |      2500      |         -          |       0        |       -        |
|        Avg        |     8.895      |       0        |     0.000      |      500       |       56.213       |       0        |     0.000      |
================================================================================================================================================

Starting benchmark: Simple value transfer

Sending to existing address

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     10.244     |      142       |     13.862     |       1        |       0.098        |   2,990,520    |  291,931.361   |
|     homestead     |     9.679      |      142       |     14.671     |       1        |       0.103        |   2,990,520    |  308,969.946   |
| tangerine-whistle |     9.173      |      142       |     15.481     |       1        |       0.109        |   2,990,520    |  326,027.236   |
|  spurious-dragon  |     9.015      |      142       |     15.752     |       1        |       0.111        |   2,990,520    |  331,733.364   |
|     byzantium     |     8.676      |      142       |     16.367     |       1        |       0.115        |   2,990,520    |  344,688.868   |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     46.786     |      710       |       -        |       5        |         -          |   14,952,600   |       -        |
|        Avg        |     9.357      |      142       |     15.175     |       1        |       0.107        |   2,990,520    |  319,593.208   |
================================================================================================================================================

Starting benchmark: Simple value transfer

Sending to non-existing address

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     9.057      |      142       |     15.678     |       1        |       0.110        |   2,982,000    |  329,233.789   |
|     homestead     |     9.862      |      142       |     14.399     |       1        |       0.101        |   2,982,000    |  302,376.458   |
| tangerine-whistle |     9.347      |      142       |     15.192     |       1        |       0.107        |   2,982,000    |  319,027.128   |
|  spurious-dragon  |     9.927      |      142       |     14.304     |       1        |       0.101        |   2,982,000    |  300,390.471   |
|     byzantium     |     8.766      |      142       |     16.199     |       1        |       0.114        |   2,982,000    |  340,183.122   |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     46.959     |      710       |       -        |       5        |         -          |   14,910,000   |       -        |
|        Avg        |     9.392      |      142       |     15.119     |       1        |       0.106        |   2,982,000    |  317,508.410   |
================================================================================================================================================

Starting benchmark: ERC20 deployment

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     7.957      |      400       |     50.271     |      100       |       12.568       |  269,966,200   | 33,928,722.212 |
|     homestead     |     7.993      |      400       |     50.043     |      100       |       12.511       |  276,366,200   | 34,575,291.495 |
| tangerine-whistle |     7.993      |      400       |     50.044     |      100       |       12.511       |  276,396,200   | 34,579,939.992 |
|  spurious-dragon  |     8.059      |      400       |     49.631     |      100       |       12.408       |  276,420,200   | 34,297,684.445 |
|     byzantium     |     8.112      |      400       |     49.312     |      100       |       12.328       |  276,420,200   | 34,077,246.258 |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     40.114     |      2000      |       -        |      500       |         -          | 1,375,569,000  |       -        |
|        Avg        |     8.023      |      400       |     49.858     |      100       |       12.464       |  275,113,800   | 34,291,480.516 |
================================================================================================================================================

################################################################################################################################################

------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |    219.022     |      3420      |       -        |      5510      |         -          | 1,405,431,600  |       -        |
|        Avg        |     8.761      |      137       |     15.615     |      220       |       25.157       |   56,217,264   | 6,416,843.310  |
================================================================================================================================================

```

Benchmark after change. 

```
______                 _                          _     
| ___ \               | |                        | |    
| |_/ / ___ _ __   ___| |__  _ __ ___   __ _ _ __| | __ 
| ___ \/ _ \ '_ \ / __| '_ \| '_ ` _ \ / _` | '__| |/ / 
| |_/ /  __/ | | | (__| | | | | | | | | (_| | |  |   <  
\____/ \___|_| |_|\___|_| |_|_| |_| |_|\__,_|_|  |_|\_\

Py-EVM/0.2.0a26/linux/cpython3.6.5

Starting benchmark: Empty block mining

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     8.310      |       0        |     0.000      |      500       |       60.170       |       0        |     0.000      |
|     homestead     |     8.546      |       0        |     0.000      |      500       |       58.510       |       0        |     0.000      |
| tangerine-whistle |     8.590      |       0        |     0.000      |      500       |       58.210       |       0        |     0.000      |
|  spurious-dragon  |     8.372      |       0        |     0.000      |      500       |       59.724       |       0        |     0.000      |
|     byzantium     |     8.012      |       0        |     0.000      |      500       |       62.408       |       0        |     0.000      |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     41.829     |       0        |       -        |      2500      |         -          |       0        |       -        |
|        Avg        |     8.366      |       0        |     0.000      |      500       |       59.768       |       0        |     0.000      |
================================================================================================================================================

Starting benchmark: Empty block import

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     8.164      |       0        |     0.000      |      500       |       61.243       |       0        |     0.000      |
|     homestead     |     8.209      |       0        |     0.000      |      500       |       60.910       |       0        |     0.000      |
| tangerine-whistle |     8.222      |       0        |     0.000      |      500       |       60.811       |       0        |     0.000      |
|  spurious-dragon  |     8.276      |       0        |     0.000      |      500       |       60.416       |       0        |     0.000      |
|     byzantium     |     8.339      |       0        |     0.000      |      500       |       59.956       |       0        |     0.000      |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     41.211     |       0        |       -        |      2500      |         -          |       0        |       -        |
|        Avg        |     8.242      |       0        |     0.000      |      500       |       60.664       |       0        |     0.000      |
================================================================================================================================================

Starting benchmark: Simple value transfer

Sending to existing address

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     7.811      |      142       |     18.179     |       1        |       0.128        |   2,990,520    |  382,848.693   |
|     homestead     |     7.816      |      142       |     18.167     |       1        |       0.128        |   2,990,520    |  382,601.529   |
| tangerine-whistle |     7.861      |      142       |     18.065     |       1        |       0.127        |   2,990,520    |  380,439.129   |
|  spurious-dragon  |     7.860      |      142       |     18.067     |       1        |       0.127        |   2,990,520    |  380,488.511   |
|     byzantium     |     7.883      |      142       |     18.013     |       1        |       0.127        |   2,990,520    |  379,354.426   |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     39.231     |      710       |       -        |       5        |         -          |   14,952,600   |       -        |
|        Avg        |     7.846      |      142       |     18.098     |       1        |       0.127        |   2,990,520    |  381,141.654   |
================================================================================================================================================

Starting benchmark: Simple value transfer

Sending to non-existing address

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     7.868      |      142       |     18.049     |       1        |       0.127        |   2,982,000    |  379,019.257   |
|     homestead     |     7.880      |      142       |     18.020     |       1        |       0.127        |   2,982,000    |  378,423.837   |
| tangerine-whistle |     7.920      |      142       |     17.930     |       1        |       0.126        |   2,982,000    |  376,519.727   |
|  spurious-dragon  |     7.954      |      142       |     17.852     |       1        |       0.126        |   2,982,000    |  374,884.830   |
|     byzantium     |     7.973      |      142       |     17.811     |       1        |       0.125        |   2,982,000    |  374,029.480   |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     39.595     |      710       |       -        |       5        |         -          |   14,910,000   |       -        |
|        Avg        |     7.919      |      142       |     17.932     |       1        |       0.126        |   2,982,000    |  376,565.477   |
================================================================================================================================================

Starting benchmark: ERC20 deployment

------------------------------------------------------------------------------------------------------------------------------------------------
|        VM         | total seconds  |    total tx    |  tx / second   |  total blocks  |  blocks / second   |   total gas    |  gas / second  |
------------------------------------------------------------------------------------------------------------------------------------------------
|     frontier      |     7.942      |      400       |     50.366     |      100       |       12.592       |  269,966,200   | 33,993,039.418 |
|     homestead     |     8.821      |      400       |     45.346     |      100       |       11.336       |  276,366,200   | 31,329,948.541 |
| tangerine-whistle |     8.935      |      400       |     44.766     |      100       |       11.192       |  276,396,200   | 30,933,067.809 |
|  spurious-dragon  |     8.688      |      400       |     46.040     |      100       |       11.510       |  276,420,200   | 31,816,017.655 |
|     byzantium     |     8.508      |      400       |     47.014     |      100       |       11.753       |  276,420,200   | 32,489,037.746 |
------------------------------------------------------------------------------------------------------------------------------------------------
|       Total       |     42.894     |      2000      |       -        |      500       |         -          | 1,375,569,000  |       -        |
…